@@ -1,8 +1,17 @@
 module github.com/dankinder/httpmock
 
-go 1.12
+go 1.19
 
 require (
-	github.com/stretchr/testify v1.3.0
+	github.com/stretchr/testify v1.8.4
 	gotest.tools/v3 v3.0.3
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/go-cmp v0.4.0 // indirect
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/objx v0.5.1 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
